abstracted table into CompletedTasks
