+++ src/routes/Classrooms.js
@@ -5,7 +5,7 @@ import TextField from '@mui/material/TextField';
 import Typography from '@mui/material/Typography';
 import { collection, onSnapshot, query, where } from "firebase/firestore";
 import React from "react";
-import ClassroomCard from '../components/classroomCard';
+import ClassroomCard from '../components/ClassroomCard';
 import { db } from '../utils/firebase';
 import { addClassroom, joinClassroom } from '../utils/mutations';
